+++ pkg/utils/nodeclaim/types.go
@@ -68,4 +68,3 @@ func IgnoreDuplicateNodeError(err error) error {
 	}
 	return nil
 }
-
